Added support for descendant taxon queries.
It's still not particularly fast, but possibly fast enough w/ caching.
The main finding here is that hard coding the id and ancestry into the
query seems to make better use of the indices than joining in those
values from the taxa table.

I've also started using Step to control flow, since it seemed
necessary when working with the node-postgres library.